requirements.txt
# SF Trending Eats Python Dependencies

# HTTP client
httpx>=0.27.0

# Database
psycopg2-binary>=2.9.9

# NLP
textblob>=0.18.0

# Google Trends (unofficial)
pytrends>=4.9.3

# Environment
python-dotenv>=1.0.1

# Math/Stats
numpy>=1.26.0

# Async support
asyncio-extras>=1.3.2
